revert: add annotations BinaryRelationWithOptionalTrigger and BinaryRelationWithEvidence, but move them into tests; move _test_reconstruct() to tests.core.test_document
