+++ tests/core/test_document.py
@@ -1,9 +1,10 @@
 import dataclasses
-from typing import List, Optional, Tuple
+import json
+from typing import Dict, List, Optional, Tuple
 
 import pytest
 
-from pytorch_ie.annotations import Span
+from pytorch_ie.annotations import Span, _post_init_single_label
 from pytorch_ie.core import Annotation
 from pytorch_ie.core.document import (
     _contains_annotation_type,
@@ -13,6 +14,16 @@ from pytorch_ie.core.document import (
     _is_optional_type,
     _is_tuple_of_annotation_types,
 )
+
+
+def _test_reconstruct(
+    annotation: Annotation, annotation_store: Optional[Dict[int, Annotation]] = None
+):
+    ann_str = json.dumps(annotation.asdict())
+    annotation_reconstructed = type(annotation).fromdict(
+        json.loads(ann_str), annotation_store=annotation_store
+    )
+    assert annotation_reconstructed == annotation
 
 
 def test_is_optional_type():
@@ -148,3 +159,105 @@ def test_get_reference_fields_and_container_types():
 
     with pytest.raises(TypeError):
         _get_reference_fields_and_container_types(Dummy)
+
+
+def test_annotation_with_optional_reference():
+    @dataclasses.dataclass(eq=True, frozen=True)
+    class BinaryRelationWithOptionalTrigger(Annotation):
+        head: Span
+        tail: Span
+        label: str
+        trigger: Optional[Span] = None
+        score: float = 1.0
+
+        def __post_init__(self) -> None:
+            _post_init_single_label(self)
+
+    head = Span(start=1, end=2)
+    tail = Span(start=3, end=4)
+    trigger = Span(start=5, end=7)
+
+    binary_relation1 = BinaryRelationWithOptionalTrigger(head=head, tail=tail, label="label1")
+    assert binary_relation1.head == head
+    assert binary_relation1.tail == tail
+    assert binary_relation1.label == "label1"
+    assert binary_relation1.score == pytest.approx(1.0)
+
+    assert binary_relation1.asdict() == {
+        "_id": binary_relation1._id,
+        "head": head._id,
+        "tail": tail._id,
+        "trigger": None,
+        "label": "label1",
+        "score": 1.0,
+    }
+
+    binary_relation2 = BinaryRelationWithOptionalTrigger(
+        head=head, tail=tail, label="label2", score=0.5, trigger=trigger
+    )
+    assert binary_relation2.head == head
+    assert binary_relation2.tail == tail
+    assert binary_relation2.trigger == trigger
+    assert binary_relation2.label == "label2"
+    assert binary_relation2.score == pytest.approx(0.5)
+
+    assert binary_relation2.asdict() == {
+        "_id": binary_relation2._id,
+        "head": head._id,
+        "tail": tail._id,
+        "trigger": trigger._id,
+        "label": "label2",
+        "score": 0.5,
+    }
+
+    annotation_store = {
+        head._id: head,
+        tail._id: tail,
+        trigger._id: trigger,
+    }
+    _test_reconstruct(binary_relation1, annotation_store=annotation_store)
+    _test_reconstruct(binary_relation2, annotation_store=annotation_store)
+
+
+def test_annotation_with_tuple_of_references():
+    @dataclasses.dataclass(eq=True, frozen=True)
+    class BinaryRelationWithEvidence(Annotation):
+        head: Span
+        tail: Span
+        label: str
+        evidence: Tuple[Span, ...]
+        score: float = 1.0
+
+        def __post_init__(self) -> None:
+            _post_init_single_label(self)
+
+    head = Span(start=1, end=2)
+    tail = Span(start=3, end=4)
+    evidence1 = Span(start=5, end=7)
+    evidence2 = Span(start=9, end=10)
+
+    relation = BinaryRelationWithEvidence(
+        head=head, tail=tail, label="label1", evidence=(evidence1, evidence2)
+    )
+    assert relation.head == head
+    assert relation.tail == tail
+    assert relation.label == "label1"
+    assert relation.score == pytest.approx(1.0)
+    assert relation.evidence == (evidence1, evidence2)
+
+    assert relation.asdict() == {
+        "_id": relation._id,
+        "head": head._id,
+        "tail": tail._id,
+        "evidence": [evidence1._id, evidence2._id],
+        "label": "label1",
+        "score": 1.0,
+    }
+
+    annotation_store = {
+        head._id: head,
+        tail._id: tail,
+        evidence1._id: evidence1,
+        evidence2._id: evidence2,
+    }
+    _test_reconstruct(relation, annotation_store=annotation_store)
